% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-and-eval.R
\name{eval.basis}
\alias{eval.basis}
\title{Evaluate the density ratio estimate}
\usage{
eval.basis(basis.fit, x.grid)
}
\arguments{
\item{basis.fit}{a list with elements documented in \code{\link{clsif}}}

\item{x.grid}{a matrix of points to evaluate the density ratio at. Note that
this should have d rows (where d is the dimension of the density/ratio space)
and each point should be a separate column.}
}
\description{
Evaluate the density ratio estimate at \code{x.grid}
}
