% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-and-eval.R
\name{fit.dr}
\alias{fit.dr}
\title{Estimate the parameters for the basis functions}
\usage{
fit.dr(x.de, x.nu, lambda, sigma.chosen, is.adaptive = FALSE, neigh.rank,
  kernel.low, kernel.high, b, fold, learning.fct)
}
\description{
This function actually does all of the work, the \code{\link{clsif}} and 
\code{\link{ulsif}} functions just call this with their respective learning
functions.
}
