% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ulsif-functions.R
\name{ulsif}
\alias{ulsif}
\title{Unconstrained Least Square Importance Fitting (ULSIF)}
\usage{
ulsif(x.de, x.nu, lambda, sigma.chosen = 0.2, is.adaptive = FALSE,
  neigh.rank = 5, kernel.low = 0.5, kernel.high = 2, b = 50,
  fold = 6)
}
\arguments{
\item{x.de}{A matrix with d rows, with one sample from p(x_de) per column.}

\item{x.nu}{A matrix with d rows, with one sample from p(x_nu) per column.}

\item{lambda}{Positive real number. Regularisation parameter, see Sugiyama,
Suzuki and Kanamori (2012) Section 6.2.1 for details}

\item{sigma.chosen}{Positive real number. Sigma for the Gaussian kernel 
radial basis functions. If this is set to zero, will be chosen via cross 
validation.}

\item{is.adaptive}{Boolean. Adaptively choose location of basis functions.}

\item{neigh.rank}{Positive integer. How many other kernels to use to compute
distance metrics.}

\item{kernel.low}{Real number. Lower bound for rescaled distances.}

\item{kernel.high}{Real number. Upper bound for rescaled distances.}

\item{b}{Positive integer. How many kernels to use.}

\item{fold}{Positive integer. How many cross validation folds to use to 
select \code{sigma.chosen}}
}
\description{
Sugiyama, Suzuki and Kanamori (2012) density ratio estimation method
with L2 penalty on the basis parameters.
}
\details{
\code{x.de} and \code{x.nu} should be the same dimension (same number of
rows), but there can an uneven number of samples (number of rows).
}
